+++ commands/music/forceskip.js
@@ -27,14 +27,11 @@ class ForceSkip extends Command {
         if (!guildEntry.hasPremiumStatus()) {
             return message.channel.createMessage(':x: Sorry but as they are resources-whores, music commands are only available to our patreon donators. Check the `bot` command for more info');
         }
-        const clientMember = message.channel.guild.members.get(client.bot.user.id);
         const connection = client.musicManager.connections.get(message.channel.guild.id);
-        if (!clientMember.voiceState.channelID || !connection || !connection.nowPlaying) {
+        if (!connection || !connection.nowPlaying) {
             return message.channel.createMessage(':x: I am not playing anything');
         }
-        const voiceChannel = message.channel.guild.channels.get(clientMember.voiceState.channelID);
-        const player = await client.musicManager.getPlayer(voiceChannel);
-        const skippedSong = await client.musicManager.skipTrack(player, connection);
+        const skippedSong = client.musicManager.skipTrack();
         return message.channel.createMessage(`:white_check_mark: Skipped **${skippedSong.info.title}**`);       
     }
 }