+++ commands/music/pause.js
@@ -27,15 +27,12 @@ class Pause extends Command {
         if (!guildEntry.hasPremiumStatus()) {
             return message.channel.createMessage(':x: Sorry but as they are resources-whores, music commands are only available to our patreon donators. Check the `bot` command for more info');
         }
-        const clientMember = message.channel.guild.members.get(client.bot.user.id);
         const connection = client.musicManager.connections.get(message.channel.guild.id);
-        if (!clientMember.voiceState.channelID || !connection || !connection.nowPlaying) {
+        if (!connection || !connection.nowPlaying) {
             return message.channel.createMessage(':x: I am not playing anything');
         }
-        const voiceChannel = message.channel.guild.channels.get(clientMember.voiceState.channelID);
-        const player = await client.musicManager.getPlayer(voiceChannel);
-        await player.setPause(player.paused ? false : true);
-        return message.channel.createMessage(`:white_check_mark: Successfully ${player.paused ? 'paused' : 'resumed'} the playback`);       
+        await connection.player.setPause(connection.player.paused ? false : true);
+        return message.channel.createMessage(`:white_check_mark: Successfully ${connection.player.paused ? 'paused' : 'resumed'} the playback`);       
     }
 }